requirements.txt
fastapi
uvicorn
pytube
paho-mqtt
python-dotenv
